Add more type of tracking + redo styles
