+++ src/components/experiments/page-react-extension/stack/CustomPragma.tsx
@@ -7,15 +7,27 @@ import * as React from "react";
 
 import { h } from "./_utils";
 
-const Paz = () => {
-  return <div>Hello h</div>;
+import { onClick } from "./on-click";
+
+const FC = () => {
+  return <button>Function</button>;
 };
 
-const Bar = React.memo(function BarInternal() {
-  return <Paz />;
+const MemoizedNamed = React.memo(function MemoizedInternal() {
+  return <button>Memoized with an internal name</button>;
+});
+const MemoizedAnonymous = React.memo(() => {
+  return <button>Memoized anonymous</button>;
 });
 
-class Foo extends React.Component {
+const ForwardedRefNamed = React.forwardRef(function ForwardedRefInternal() {
+  return <button>ForwardedRef with an internal name</button>;
+});
+const ForwardedRefAnonymous = React.forwardRef(() => {
+  return <button>ForwardedRef anonymous</button>;
+});
+
+class Class extends React.Component<any> {
   override render() {
     console.log(
       "[CUSTOM] Test component render log in all browser dev tools",
@@ -23,15 +35,35 @@ class Foo extends React.Component {
     );
     console.trace();
     // console.log("Class current fiber:", this._reactInternals);
-    return (
-      <div>
-        <Bar />
-        <Paz />
-      </div>
-    );
+    return <button>Class</button>;
   }
 }
 
 export const CustomPragma = () => {
-  return <Foo />;
+  return (
+    <fieldset
+      style={{ display: "inline", border: "1px dashed blue" }}
+      onClick={onClick}
+    >
+      <legend>Click Tracker</legend>
+
+      {/* Render FC twice, to see how we work with same element twice */}
+      <FC />
+      <FC />
+
+      <hr />
+
+      <Class with={{ props: true }} />
+
+      <hr />
+
+      <MemoizedNamed />
+      <MemoizedAnonymous />
+
+      <hr />
+
+      <ForwardedRefNamed />
+      <ForwardedRefAnonymous />
+    </fieldset>
+  );
 };
